+++ src/components/template-components/Elements/Formatter/ImageFormatter.tsx
@@ -4,12 +4,14 @@ import React from 'react';
 import { getImageThumbnailUrl } from '../../../../utils/template-utils/utils';
 
 const ImageFormatter: React.FC<IImageProps> = ({ value }) => {
+  console.log({ value });
   const image = value[0];
-  const url = getImageThumbnailUrl(image, 256);
 
+  const url = getImageThumbnailUrl(image, 256);
+  console.log({ url });
   return (
     <div className={styles.imageItem}>
-      <img style={{ width: '20px', height: '20px' }} src={url} alt="" />
+      <img src={url} alt={value[0]} style={{ width: '50px', marginRight: '10px' }} />{' '}
     </div>
   );
 };
