+++ src/utils/template-utils/constants/index.ts
@@ -83,6 +83,67 @@ const DEFAULT_SELECTED_PRESET: IActivePresetSettings = {
   selectedView: DEFAULT_SELECT_OPTION,
 };
 
+const PREVIEWER = 'previewer';
+const ADDITION = 'addition';
+
+const FILEEXT_ICON_MAP: any = {
+  // text file
+  md: 'txt.png',
+  txt: 'txt.png',
+
+  // pdf file
+  pdf: 'pdf.png',
+
+  // document file
+  doc: 'word.png',
+  docx: 'word.png',
+  odt: 'word.png',
+  fodt: 'word.png',
+
+  ppt: 'ppt.png',
+  pptx: 'ppt.png',
+  odp: 'ppt.png',
+  fodp: 'ppt.png',
+
+  xls: 'excel.png',
+  xlsx: 'excel.png',
+  ods: 'excel.png',
+  fods: 'excel.png',
+
+  // video
+  mp4: 'video.png',
+  ogv: 'video.png',
+  webm: 'video.png',
+  mov: 'video.png',
+  flv: 'video.png',
+  wmv: 'video.png',
+  rmvb: 'video.png',
+
+  // music file
+  mp3: 'music.png',
+  oga: 'music.png',
+  ogg: 'music.png',
+  flac: 'music.png',
+  aac: 'music.png',
+  ac3: 'music.png',
+  wma: 'music.png',
+
+  // image file
+  jpg: 'pic.png',
+  jpeg: 'pic.png',
+  png: 'pic.png',
+  svg: 'pic.png',
+  gif: 'pic.png',
+  bmp: 'pic.png',
+  ico: 'pic.png',
+
+  // folder dir
+  folder: 'folder-192.png',
+
+  // default
+  default: 'file.png',
+};
+
 // Exported Constants
 export {
   POSSIBLE,
@@ -98,4 +159,7 @@ export {
   INITIAL_CURRENT_STATE,
   DEFAULT_SELECTED_PRESET,
   DEFAULT_SELECT_OPTION,
+  FILEEXT_ICON_MAP,
+  PREVIEWER,
+  ADDITION,
 };
